requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Vision-Language Models
transformers>=4.35.0
pillow>=10.0.0
timm>=0.9.0

# LLM Optimization
peft>=0.6.0
bitsandbytes>=0.41.0
accelerate>=0.24.0

# Knowledge Retrieval
sentence-transformers>=2.2.0
faiss-gpu>=1.7.4

# NLP Tools
spacy>=3.7.0
nltk>=3.8.0
rouge-score>=0.1.2

# APIs and Data
requests>=2.31.0
wikipedia-api>=0.6.0
h5py>=3.9.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0.0
tqdm>=4.66.0
tensorboard>=2.14.0
wandb>=0.15.0

# Additional
scikit-learn>=1.3.0
